requirements.txt
amqp==5.1.1
aniso8601==9.0.1
async-timeout==4.0.2
bcrypt==4.0.1
billiard==4.1.0
bleach==6.0.0
blinker==1.6.2
cachetools==5.3.1
celery==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.0.7
cycler==0.11.0
dnspython==2.3.0
email-validator==2.0.0.post2
Flask==2.3.2
Flask-Login==0.6.2
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Security-Too==5.2.0
Flask-SQLAlchemy==3.0.5
Flask-SSE==1.0.0
Flask-WTF==1.1.1
fonttools==4.39.4
gevent==22.10.2
google-api-core==2.11.1
google-api-python-client==2.91.0
google-auth==2.21.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.1
greenlet==2.0.2
httplib2==0.22.0
idna==3.4
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
kombu==5.3.1
MarkupSafe==2.1.2
matplotlib==3.7.1
numpy==1.24.3
oauthlib==3.2.2
packaging==23.1
passlib==1.7.4
Pillow==9.5.0
prompt-toolkit==3.0.38
protobuf==4.23.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
python-dateutil==2.8.2
pytz==2023.3
redis==4.6.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.15
typing_extensions==4.6.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
vine==5.0.0
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==2.3.4
WTForms==3.0.1
zope.event==5.0
zope.interface==6.0
httplib2==0.22.0
pyparsing==3.1.1
